bt-160: * mock vars, fix nothing found width on small screen width
