+++ frontend/src/bundles/chat/components/chat-list/chat-list-search/styles.module.scss
@@ -19,5 +19,5 @@
 
 .search:hover,
 .search:focus {
-    outline: 1px solid var(--st-tropaz);
+    outline: 2px solid var(--st-tropaz);
 }
